live: Reuse --record option to keep recorded data
The live command is used to record the data in a temporary directory and
discards it after the execution.

Someone might want to see the replay output immediately, but also wants
to keep the recorded data to run another analysis command.

The current --record option is only used for script command, but can
also be used for live command for this purpose.

  $ rm -rf uftrace.data
  $ uftrace --record -F main t-abc
  # DURATION     TID     FUNCTION
              [109070] | main() {
              [109070] |   a() {
              [109070] |     b() {
              [109070] |       c() {
     0.933 us [109070] |         getpid();
     3.617 us [109070] |       } /* c */
     4.207 us [109070] |     } /* b */
     4.844 us [109070] |   } /* a */
     9.650 us [109070] | } /* main */

  $ uftrace replay
  # DURATION     TID     FUNCTION
              [109070] | main() {
              [109070] |   a() {
              [109070] |     b() {
              [109070] |       c() {
     0.933 us [109070] |         getpid();
     3.617 us [109070] |       } /* c */
     4.207 us [109070] |     } /* b */
     4.844 us [109070] |   } /* a */
     9.650 us [109070] | } /* main */

Closes: #1038

+++ cmds/live.c
@@ -104,30 +104,32 @@ int command_live(int argc, char *argv[], struct opts *opts)
 	};
 	int ret;
 
-	tmp_dirname = template;
-	umask(022);
-	fd = mkstemp(template);
-	if (fd < 0) {
-		if (errno != EPERM)
-			pr_err("cannot access to /tmp");
+	if (!opts->record) {
+		tmp_dirname = template;
+		umask(022);
+		fd = mkstemp(template);
+		if (fd < 0) {
+			if (errno != EPERM)
+				pr_err("cannot access to /tmp");
 
-		fd = mkstemp(template + sizeof("/tmp/") - 1);
+			fd = mkstemp(template + sizeof("/tmp/") - 1);
 
-		if (fd < 0)
-			pr_err("cannot create temp name");
-		tmp_dirname += sizeof("/tmp/") - 1;
+			if (fd < 0)
+				pr_err("cannot create temp name");
+			tmp_dirname += sizeof("/tmp/") - 1;
+		}
+
+		close(fd);
+		unlink(tmp_dirname);
+
+		atexit(cleanup_tempdir);
+
+		sa.sa_handler = sigsegv_handler;
+		sigfillset(&sa.sa_mask);
+		sigaction(SIGSEGV, &sa, NULL);
+
+		opts->dirname = tmp_dirname;
 	}
-
-	close(fd);
-	unlink(tmp_dirname);
-
-	atexit(cleanup_tempdir);
-
-	sa.sa_handler = sigsegv_handler;
-	sigfillset(&sa.sa_mask);
-	sigaction(SIGSEGV, &sa, NULL);
-
-	opts->dirname = tmp_dirname;
 
 	if (opts->list_event) {
 		if (geteuid() == 0)
